It works!
User can enter a label in the input field in the create label menu. It will then show up in the dropdown in the select label menu. Saves locally. Delete label also works. Select label and back buttons are supposed to go to training scene but the build settings changed so there is no training scene.