Clinician can modify base size of displayed training image
StimulusManager.cs updated with methods to modify base size.

SetBaseSize() controls the image size.
ModifyBaseSizeWithSlider() takes in the slider values and calles SetBaseSize()
ResetBaseSize() resets the base size to default

ResetSPO(), which is called when a new image is selected from the inventory, no longer resets the base size. This is now handled through the UI icon to reset the image's base size.