Updated scripts for controlling training
Renamed StartTraining.cs to TrainingController.cs

Added a method to interrupt the countdown. While it doesn't stop training, it does allow the Clinician to cancel the countdown.

Exposed a variable `trainingLengthSeconds`  to control length of training window in UI, if we so desire. Set default to 8s.

Added a method to display a small UI element in the bottom left corner to indicate the time left in the training window. We can comment this out if we don't see the value.