TrainingPageManager updates number of windows in BCIController when the a tab is initialized or activated
On Start the method InitalizeViews() method calls UpdateTrialLength(). While this updated the UI settings (such as how long the training runs for), the values on the BCI Controller weren't being updated.

The fix was to move the lines from TrialLengthChanged() that updated the BCIController to UpdateTrialLength(), as TrialLengthChanged() calls UpdateTrialLength() anyways.

Thus, IntializeViews() on tab Start() will ensure the number of windows in BCIController is up to date with the UI value.

Then added an OnEnable() method that also calls InitializeViews() whenever the tab is re-enabled